Move CollateralType to useCollateralTypes
