+++ v3/ui/src/components/accounts/Position/Manage/MaintainCRatio.tsx
@@ -3,7 +3,7 @@ import { Deposit } from './Deposit';
 import { Alert, AlertIcon, Box, Heading } from '@chakra-ui/react';
 import { formatValue } from '@snx-v3/format';
 import { FC } from 'react';
-import { CollateralType } from '../../../../utils/types';
+import { CollateralType } from '@snx-v3/useCollateralTypes';
 
 interface Props {
   collateral: CollateralType;
