+++ v3/ui/src/components/accounts/Position/Manage/Withdraw.tsx
@@ -1,7 +1,7 @@
 import { Text, Box, Flex, Heading } from '@chakra-ui/react';
 import { FC } from 'react';
 import { parseUnits } from '@snx-v3/format';
-import { CollateralType } from '../../../../utils/types';
+import { CollateralType } from '@snx-v3/useCollateralTypes';
 import { Balance } from '../../Deposit/Balance';
 import { NumberInput } from './NumberInput';
