+++ v3/ui/src/components/accounts/Position/Rewards/RewardsDistributorTitle.tsx
@@ -4,7 +4,7 @@ import { FC, useCallback, useState } from 'react';
 import { useContractRead, useContractWrite, useToken } from 'wagmi';
 import { useContract, useSnxProxy } from '../../../../hooks/useContract';
 import { contracts } from '../../../../utils/constants';
-import { CollateralType } from '../../../../utils/types';
+import { CollateralType } from '@snx-v3/useCollateralTypes';
 import { Amount } from '../../../shared/Amount/Amount';
 import { RewardRate } from './RewardRate';
 import { RewardsDistributor } from './RewardsDistributor';
