+++ v3/ui/src/utils/types.ts
@@ -1,17 +1,5 @@
+import { CollateralType } from '@snx-v3/useCollateralTypes';
 import { BigNumber } from 'ethers';
-
-export type CollateralType = {
-  depositingEnabled: boolean;
-  issuanceRatioD18: BigNumber;
-  liquidationRatioD18: BigNumber;
-  liquidationRewardD18: BigNumber;
-  minDelegationD18: BigNumber;
-  oracleNodeId: string;
-  tokenAddress: `0x${string}`;
-  symbol: string;
-  price?: BigNumber;
-  logo: string;
-};
 
 export type LiquidityPositionType = {
   id: string;
